refactor(workspace): split WorkspaceClient into focused modules
- Extract ProcessEntry type and sendToEntryWindows utility to types.ts
- Extract process pool lifecycle (entries, windowToProject, loadProject,
  prewarmProject, eviction, dormant cap, direct ports, restart) to
  WorkspaceHostPool (~400 lines)
- Extract host event routing (11 event-type switch, rate-limit guard,
  toast broadcasts) to WorkspaceHostEventRouter (~180 lines)
- Extract CopyTree operations (generateContext, testConfig, cancelContext,
  cancelAllContext) to WorkspaceCopyTreeClient (~120 lines)
- WorkspaceClient.ts remains as public facade (~430 lines, down from 1314)
- All public method signatures, EventEmitter emits, and singleton exports
  unchanged
- 68/68 existing tests pass with zero assertion changes
