+++ electron/services/workspace-client/index.ts
@@ -1,0 +1,16 @@
+export { type ProcessEntry, type CopyTreeProgressCallback, sendToEntryWindows } from "./types.js";
+export {
+  WorkspaceHostPool,
+  type RouteHostEventFn,
+  type EmitFn as PoolEmitFn,
+  type WorkspaceHostPoolDeps,
+} from "./WorkspaceHostPool.js";
+export {
+  WorkspaceHostEventRouter,
+  type EmitFn as EventRouterEmitFn,
+  type WorkspaceHostEventRouterDeps,
+} from "./WorkspaceHostEventRouter.js";
+export {
+  WorkspaceCopyTreeClient,
+  type WorkspaceCopyTreeClientDeps,
+} from "./WorkspaceCopyTreeClient.js";
